% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{binomConfint}
\alias{binomConfint}
\title{binomConfint}
\source{
swirl.com
}
\usage{
binomConfint(p_prime = "sample proportion", n = "number of observations",
  conf = ".9, .95, .99... etc")
}
\arguments{
\item{p_prime}{= "sample proportion"
n = "number of observations"
conf = confidence interval}
}
\description{
binomConfint
}
\examples{
binomConfint(.6, 100, .95)
}
\author{
swirl.com
}

