% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{completest.cases}
\alias{completest.cases}
\title{completest.cases}
\usage{
completest.cases(theDataframe)
}
\description{
returns rows without NAs
}
\examples{
# Remove na from stockQuoteBig
stocks <- stockQuoteBig(c("AAPL"), 2012)
aaplNoNA <- completest.cases(stocks$AAPL$byDay)
}

