% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{d_beta}
\alias{d_beta}
\title{Stock beta generator}
\usage{
d_beta(Stock, Index)
}
\arguments{
\item{stock}{A stock ticker,i.e, "AAPL"}

\item{index}{A stock index, i.e, "SPY"}
}
\value{
beta for stock
}
\description{
Stock beta generator
}
\examples{
d_beta("AAPL", "SPY")
}

