% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{expandNum}
\alias{expandNum}
\title{Expands columns with categorical variables into ones and zeroes}
\usage{
expandNum(dataFrame = "A data.frame", colNum = "c(a,b,c,d...)", ...)
}
\arguments{
\item{dataFrame}{Where the data is stored}

\item{colNum}{Columns to be expanded}
}
\description{
Expands columns with categorical variables into ones and zeroes
}
\examples{
expandNum(animals, c(4,5,6,7))
}

