% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{firstDayMonth}
\alias{firstDayMonth}
\title{firstDayMonth}
\usage{
firstDayMonth(x)
}
\arguments{
\item{x}{= list of dates}
}
\description{
Returns the first date of each month in list of dates
}
\examples{

}

