% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{firstDayYear}
\alias{firstDayYear}
\title{firstDayYear}
\usage{
firstDayYear(x)
}
\arguments{
\item{x}{= list of dates}
}
\description{
Returns the first date of each year in list of dates
}
\examples{

}

