% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{lehmer}
\alias{lehmer}
\title{Lehmer}
\usage{
lehmer(n = "number of columns and rows, nXn matrix",
  roundTo = "rounding of matrix if desired")
}
\arguments{
\item{n}{= number of rows/columns}
}
\description{
Lehmer
}
\examples{
lehmer(10)
}

