% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{meanplot}
\alias{meanplot}
\title{meanplot}
\usage{
meanplot(dataset, columns, sens)
}
\arguments{
\item{dataset}{= a dataset
columns = the two columns to be plotted
sens = integer}
}
\description{
plots the average of points based on distance from one another given by sens
}
\examples{
Everything below should run together.
x <- -10:10
y <- jitter(-(-10:10)*10 + (-10:10)^2 + (-10:10)^3, 500)
xy <- data.frame("x" = x, "y" = y)
meanplot(xy, c(1,2), sens = 1)
for(i in 1:3) {
meanplot(xy, c(1,2), sens = i)
}
}

