% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{packageLoaded}
\alias{packageLoaded}
\title{Checks if a package is loaded}
\usage{
packageLoaded(name)
}
\arguments{
\item{a}{package name}
}
\description{
Checks if a package is loaded
}
\examples{
packageLoaded("ggplot2")
}

