% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stockData}
\alias{stockData}
\title{stckData}
\usage{
stockData(stockVector, from, byDay = FALSE, byYear = FALSE,
  byMonth = FALSE)
}
\description{
Returns data about selected stocks
}
\examples{
stock <- stockData(c("AAPL", "AKRX", "SPY"), 2009)
}

