% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stockQuoteBig}
\alias{stockQuoteBig}
\title{stockQuoteBig}
\usage{
stockQuoteBig(stockName = "Ex. \\"AAPL\\" ", year)
}
\arguments{
\item{stock}{= stockTicker}
}
\description{
stockQuoteBig
}
\examples{

aapl <- stockQuoteBig("AAPL", "AKRX")
aapl$byDay
aapl$byMonth
aapl$byYear$Date

# Correlation between sh (S&P inverse) and S&P ETF
sh  <- stockQuoteBig("SH", 2010)
spy  <- stockQuoteBig("SPY", 2010)
cor(spy$byMonth$Percent.Change[1:80], sh$byMonth$Percent.Change[1:80])
}

