% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tStatistic}
\alias{tStatistic}
\title{tStatistic}
\usage{
tStatistic(sample, nullHyp, stDev)
}
\arguments{
\item{sample}{= sampled data
nullHyp = null hypothesis mean
stDev = standard deviation (ONLY IF KNOWN)}
}
\description{
tStatistic
}

