% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tTest}
\alias{tTest}
\title{tTest}
\usage{
tTest(sample, confidence, stDev)
}
\arguments{
\item{sample}{= sampled data
confidence = confidence of choice, usually .95
stDev = standard deviaion (ONLY IF KNOWN)}
}
\description{
tTest
}
\examples{
students <- rnorm(10000,mean = 70000, sd = 15000)
sampleStud <- sample(students,size = 60,replace = FALSE)
tTest(sampleStud, .95)
}
\seealso{
t.test(sampleStud)
}

