% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tTestDifference}
\alias{tTestDifference}
\title{tTestDifference}
\usage{
tTestDifference(group1, group2, confidence)
}
\description{
tTestDifference
}
\examples{
diffTInv(g1, g2, .95)
}
\seealso{
t.test(difference)
}

