% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{upperTriangular}
\alias{upperTriangular}
\title{Converts a matrix to upper triangular form}
\usage{
upperTriangular(A, roundedTo = 4)
}
\arguments{
\item{A}{= matrix}

\item{roundedTo}{= integer to round to}
}
\description{
Converts a matrix to upper triangular form
}
\examples{
 K <- matrix(c(1,4,5,3,5,
             6,5,4,3,2,
             6,7,8,9,0,
             4,3,2,5,6,
             6,6,7,8,8), nrow=5, byrow=TRUE)

upperTriangular(K)
}

