% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{urlTables}
\alias{urlTables}
\title{Downloads tables from url}
\usage{
urlTables(url, ...)
}
\arguments{
\item{url}{A url}
}
\description{
Downloads tables from url
}
\examples{
starTable <- urlTables("https://en.wikipedia.org/wiki/Star")
# starTable[[3]]
}

