% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{yahooFin}
\alias{yahooFin}
\title{Gets stock data that you might want}
\usage{
yahooFin(stock)
}
\arguments{
\item{a}{stock symbol}
}
\description{
Gets stock data that you might want
}
\examples{
AAPL <- yahooFin("AAPL")
}

